Build MMIX on GNU/Linux.
In October 2006, I started to build MMIX on GNU/Linux in an automated
fashion. The SPEC file mmix.spec was first created for RPM on OpenSUSE,
but works just as well on Kubuntu with debbuild.

The latest source drop can be created using ‘git archive’ and the two
changefiles are available AS IS.