Run some basic tests.
These are described in the README file and executed by the cygwin port:
https://github.com/cygwinports-extras/mmixware/blob/master/mmixware.cygport#L45

We generate a POSIX-conforming test script.

'echo -e "10000\nq"' (for support of the '\n' escape sequence) works in
'bash' but not in 'dash' (the default '/bin/sh'ell in Ubuntu):
https://unix.stackexchange.com/questions/65803/why-is-printf-better-than-echo
Using 'printf' instead works with the default setting of '%_buildshell'.

The second test from the README file:

   $> ./mmixal silly.mms
   $> ./mmix -i silly
   mmix> i silly.run

actually produces identical output as in 'silly.out', when copy-pasted
from the shell window. However, the two 'Warnings' are printed on
'stderr' and appear asynchronously in 'silly.new' when invoked like

   $> echo "i silly.run" | ./mmix -i silly > silly.new 2>&1

So, this initial test is only performed manually.