DRY up the Makefile.
The rules for mmix-pipe and mmix-sim are similar. With the
.SECONDEXPANSION feature of GNU make, this can be made clear.

And we can replace 'abstime' by 'perl/date'. No need to work with
the temporary file 'abstime.h' at all. Note: 'perl' is used because
'sed' on Linux and BSD (macos) have incompatible invocation options.

This brings ideas from branches 'make-mmix-sim.o' and 'secondexpansion'
into production.