More conservative sed invocation.
sed on Linux and sed on MacOS are not quite identical in their
respective interpretation of the '-i' (--in-place) option. MacOS
requires a separate parameter, which may be the empty string "". Linux
takes 'the stuff after -i' as the backup suffix, so plain '-i'
suppresses the creation of backups for in-place-modified files.

The '-e' option is required by MacOS sed and optional for Linux sed.