Put l and g into index.
Both these global variables are defined in mmix-pipe.w$86. Because of
their nature as single lowercase characters, CWEB does _not_ put their
uses into the index.

However, because of their quite central role for the workings of the
MMIX system, we use the present 'kludge' to make them appear (with
amendments) in the indexes of mmix-config.w and mmmix.w.

Adding similar entries to the index of mmix-pipe.w itself is left as an
exercise for the reader.