Fix bracketing for MMIXAL§94.
Section 94 is not surrounded by parentheses, so only the first line is
actually associated with the 'else' in section 86; the rest of the code
in section 94 is executed unconditionally.

Using a 'switch' statement in the second 'else if' as in the final
'else' in section 86, fixes this and also simplifies the overall
bracketing situation.